feat: endpoint to confirm booking
